hsCoefAnaRec

recursive version of regression coefficient analysis

hsCoefAnaRec(
  data,
  evtNums = unique(data$evtID),
  combi = NULL,
  tree = TRUE,
  resframe = NULL,
  uselm = FALSE,
  ...,
  dbg.level = 1
)

Arguments

data

data frame containing columns tstamp (time stamp), pval (probe value), lval (lab value), evtID (event ID)

evtNums

event numbers to be considered for the analysis (default: all distinct values provided in column evtID of data)

combi

current combination to be evaluated and to be the base for the next combinations to be determined

tree

if TRUE, result is given in a tree structure, otherwise as a data frame

resframe

if tree is FALSE, this argument contains the results that have been found so far in a data frame

uselm

if TRUE, the lm function is used to calculate the linear regression, otherwise (uselm == FALSE) the regression is calculated "manually" which is much faster. default: FALSE

...

further arguments passed to hsCombiLinReg, e.g. clever

dbg.level

debug level

Value

Recursive list representing a tree structure. At the top level the list contains elements e<i> where <i> are the event IDs to be considered (elements in evtNums). The sub lists below the top level (but not the "leafs" of the tree) also contain elements e<j> where <j> are the "remaining" event IDs, i.e. the IDs that do not yet occur in the "path" of event IDs leading to the respective sub tree. These sub lists also have elements combi (vector of event IDs representing the respective event combination) and linreg containing the results from linear regression. In fact, linreg is a data frame with each line representing the slope and offset of the linear regression through np number of points, taken from the events in combi.