Data availability of time series data

Data availability of time series data

hsDataAvailability.old(
  info,
  dateFirst = NULL,
  dateLast = NULL,
  tstep = NULL,
  dbg = FALSE
)

Arguments

info

list with the named elements (mdb: full path to Access database, tbl: table name, tsField: name of timestamp field, parField: name of parameter field)

dateFirst

Date object representing first date to be considered

dateLast

Date object representing last date to be considered

tstep

expected time step between time stamps in seconds. Default: minimum time difference found between consecutive timestamps in given interval

dbg

If TRUE, debug messages will be shown

Value

data.frame with each row representing a day within the specified time interval and columns \(intervalBeg\) (day), \(n<Par>\) (number of non-NA-values in column <Par> within the interval) and \(p<Par>\) (data availability of parameter <Par> in percent = number of available non-NA-values divided by maximum possible number of non-NA-values per day (= 86400 / tstep).